Locate the original name and location of an object

Locate the original name and location of an object

find_object_by_value(
  object,
  from = parent.frame(),
  until = emptyenv(),
  which = c("first", "last", "all")
)

Arguments

object

The R object whose location should be identified.

from

An environment, or an object with an environment (e.g. a function and a formula), to start search from.

until

An environment to consider the last parent environment. If until is not one of the parent environments, then emptyenv() is the last one. It is also possible to specify a list of alternative environments.

which

If "first" or "last", then the first or the last occurance of object among the parent frames is identified and returned. If "all", then all occurances are returned.

Value

If which = "first" or which = "last", then a named list with elements name and envir, where name is the name of object

as it is named in environment envir, i.e. identical(envir[[name]], object). If which = "all", then a list of (name, environment) lists are returned; one for each matching occurence. If the object could not be located when searching from environment from, then NULL is returned.

Examples

find_object_by_value(rnorm)
#> $name
#> [1] "rnorm"
#> 
#> $envir
#> <environment: package:stats>
#> attr(,"name")
#> [1] "package:stats"
#> attr(,"path")
#> [1] "/home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/stats"
#> 

my_sum <- rnorm
find_object_by_value(my_sum)
#> $name
#> [1] "my_sum"
#> 
#> $envir
#> <environment: 0x5653f06913d8>
#> 

find_object_by_value(my_sum, from = parent_env())
#> $name
#> [1] "rnorm"
#> 
#> $envir
#> <environment: package:stats>
#> attr(,"name")
#> [1] "package:stats"
#> attr(,"path")
#> [1] "/home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/stats"
#> 

my_sum <- local(sum)
find_object_by_value(my_sum)
#> $name
#> [1] "my_sum"
#> 
#> $envir
#> <environment: 0x5653f06913d8>
#> 

my_find <- function(object, envir = parent.frame()) {
  find_object_by_value(object, from = envir)
}

my_fcn <- local({
  g <- sum  ## this is never found
  function(f = NULL) {
    if (is.null(f)) {
      f <- g  ## this is found first
      my_find(f)
    } else {
      my_find(f, envir = parent.frame()) # skip argument 'f'
    }
  }
})
my_fcn()
#> $name
#> [1] "f"
#> 
#> $envir
#> <environment: 0x5653f0c3bfb8>
#> 
my_fcn(my_sum)
#> $name
#> [1] "my_sum"
#> 
#> $envir
#> <environment: 0x5653f06913d8>
#> 

topenv()
#> <environment: R_GlobalEnv>