Find all parent environments of an environment

Find all parent environments of an environment

parent_envs(envir = parent.frame(), until = emptyenv(), extra = 0L)

Arguments

envir

An environment or an object with an environment, e.g. a function or a formula.

until

A environment to consider the last parent environment. If until is not one of the parent environments, then emptyenv() is the last one. It is also possible to specify a list of alternative environments.

extra

Maximum number of additional parent environments to include after a matching "until" one was identified.

Value

A named list of environments, where the first element is envir and the last is until or emptyenv().

Examples

parent_envs(emptyenv())
#> $R_EmptyEnv
#> <environment: R_EmptyEnv>
#> 
parent_envs(baseenv())
#> $base
#> <environment: base>
#> 
#> $R_EmptyEnv
#> <environment: R_EmptyEnv>
#> 
parent_envs(globalenv())
#> $R_GlobalEnv
#> <environment: R_GlobalEnv>
#> 
#> $`package:environments`
#> <environment: package:environments>
#> attr(,"name")
#> [1] "package:environments"
#> attr(,"path")
#> [1] "/tmp/hb/Rtmp7MatSw/temp_libpath2565a966fbdf2c/environments"
#> 
#> $`package:stats`
#> <environment: package:stats>
#> attr(,"name")
#> [1] "package:stats"
#> attr(,"path")
#> [1] "/home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/stats"
#> 
#> $`package:graphics`
#> <environment: package:graphics>
#> attr(,"name")
#> [1] "package:graphics"
#> attr(,"path")
#> [1] "/home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/graphics"
#> 
#> $`package:grDevices`
#> <environment: package:grDevices>
#> attr(,"name")
#> [1] "package:grDevices"
#> attr(,"path")
#> [1] "/home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/grDevices"
#> 
#> $`package:utils`
#> <environment: package:utils>
#> attr(,"name")
#> [1] "package:utils"
#> attr(,"path")
#> [1] "/home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/utils"
#> 
#> $`package:datasets`
#> <environment: package:datasets>
#> attr(,"name")
#> [1] "package:datasets"
#> attr(,"path")
#> [1] "/home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/datasets"
#> 
#> $`package:methods`
#> <environment: package:methods>
#> attr(,"name")
#> [1] "package:methods"
#> attr(,"path")
#> [1] "/home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/methods"
#> 
#> $Autoloads
#> <environment: 0x5567c08e03e0>
#> attr(,"name")
#> [1] "Autoloads"
#> 
#> $`tools:callr`
#> <environment: 0x5567c1689090>
#> attr(,"name")
#> [1] "tools:callr"
#> 
#> $base
#> <environment: base>
#> 
#> $R_EmptyEnv
#> <environment: R_EmptyEnv>
#> 
parent_envs(new.env(parent = baseenv()))
#> $`0x5567ce5a1670`
#> <environment: 0x5567ce5a1670>
#> 
#> $base
#> <environment: base>
#> 
#> $R_EmptyEnv
#> <environment: R_EmptyEnv>
#> 

f <- local({
  a <- 42
  local({
    pi <- 3.14
    function() pi * a
  })
})
f_envs <- parent_envs(f, until = environment())
names(f_envs)
#> [1] "0x5567ce60dab8" "0x5567ce60ec38" "0x5567ce44bfb8"

f_envs <- parent_envs(f, until = environment(), extra = 1L)
names(f_envs)
#> [1] "0x5567ce60dab8" "0x5567ce60ec38" "0x5567ce44bfb8" "0x5567ce22b6b0"